Move AdFormat.ts and AdViewPosition.ts into AdViewProps.ts
