--- src/types/AdViewPosition.ts
@@ -1,30 +1,0 @@
-import { NativeModules } from "react-native";
-
-const { AppLovinMAX } = NativeModules;
-
-const {
-    TOP_CENTER_POSITION,
-    TOP_LEFT_POSITION,
-    TOP_RIGHT_POSITION,
-    CENTERED_POSITION,
-    CENTER_LEFT_POSITION,
-    CENTER_RIGHT_POSITION,
-    BOTTOM_LEFT_POSITION,
-    BOTTOM_CENTER_POSITION,
-    BOTTOM_RIGHT_POSITION,
-} = AppLovinMAX.getConstants();
-
-/**
- * Defines a position of a banner and MRec ad.
- */
-export enum AdViewPosition {
-    TOP_CENTER = TOP_CENTER_POSITION,
-    TOP_LEFT = TOP_LEFT_POSITION,
-    TOP_RIGHT = TOP_RIGHT_POSITION,
-    CENTERED = CENTERED_POSITION,
-    CENTER_LEFT = CENTER_LEFT_POSITION,
-    CENTER_RIGHT = CENTER_RIGHT_POSITION,
-    BOTTOM_LEFT = BOTTOM_LEFT_POSITION,
-    BOTTOM_CENTER = BOTTOM_CENTER_POSITION,
-    BOTTOM_RIGHT = BOTTOM_RIGHT_POSITION,
-}
