+++ src/types/index.ts
@@ -1,7 +1,5 @@
 export * from './AppLovinMAX';
 export * from './TargetingData';
 export * from './AdInfo';
-export * from './AdFormat';
 export * from './AdViewProps';
-export * from './AdViewPosition';
 export * from './NativeAdViewProps';
